requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
argcomplete==3.4.0
async-timeout==4.0.3
attrs==23.2.0
boto==2.49.0
cachetools==5.4.0
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==2.1.1
clarabel==0.9.0
cloudpickle==3.0.0
contourpy==1.1.1
crcmod==1.7
cryptography==42.0.8
cvxpy==1.5.2
cycler==0.12.1
ecos==2.0.14
fasteners==0.19
filelock==3.13.1
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.2.0
gcs-oauth2-boto-plugin==3.2
google-apitools==0.5.32
google-auth==2.17.0
google-auth-httplib2==0.2.0
google-reauth==0.1.1
gsutil==5.30
gym==0.26.2
gym-notices==0.0.8
httplib2==0.20.4
idna==3.4
imageio==2.34.2
importlib_metadata==8.0.0
importlib_resources==6.4.0
Jinja2==3.1.3
joblib==1.4.2
kiwisolver==1.4.5
lazy_loader==0.4
learn2learn==0.2.0
lmdb==1.5.1
MarkupSafe==2.1.5
matplotlib==3.7.5
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
networkx==3.0
numpy==1.24.1
oauth2client==4.1.3
opencv-python==4.10.0.84
osqp==0.6.7.post0
packaging==24.1
pandas==2.0.3
pillow==10.2.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyOpenSSL==24.1.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
pyu2f==0.1.5
PyWavelets==1.4.1
qdldl==0.1.7.post4
qpth==0.0.17
requests==2.28.1
retry-decorator==1.1.1
rsa==4.7.2
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
scs==3.2.6
seaborn==0.13.2
six==1.16.0
sympy==1.12
tabulate==0.9.0
threadpoolctl==3.5.0
tifffile==2023.7.10
torch==1.13.0+cu116
torchaudio==0.13.0+cu116
torchvision==0.14.0+cu116
tqdm==4.66.4
triton==2.1.0
typing_extensions==4.9.0
tzdata==2024.1
urllib3==1.26.13
yarl==1.9.4
zipp==3.19.2
